requirements.txt
docling-core
pandas~=2.3.0
# Comment or remove package torch if you want to use the Dockerfile and install just the CPU option
torch~=2.7.1
docling~=2.39.0
pdf2image~=1.17.0
transformers~=4.53.0
protobuf~=6.31.1
streamlit~=1.46.1
# Comment or remove package accelerate if you want to use the Dockerfile and install just the CPU option
accelerate~=1.8.1
langchain~=0.3.26
langchain-ollama~=0.3.4
easyocr~=1.7.2
pydantic~=2.11.7
langchain-core~=0.3.68
langchain-openai~=0.3.28
python-dotenv~=1.1.1
langchain-google-genai~=2.1.9
numpy~=2.2.6
